Introduce a global buffer for extra arrays.
Many patterns need some extra storage space. Rather than allocating
multiple small buffers, this approach is to allocate one big global
one and then re-use parts of it for individual patterns.

The code is uglier and more error-prone, but that's the trade-off due
to the limited memory available on the platform.